backend/pyproject.toml
[project]
name = "helloagents-deep-researcher"
version = "0.0.1"
description = "Fully local web research and summarization assistant powered by HelloAgents."
authors = [
    { name = "Lance Martin" }
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic",
    "python-dotenv==1.0.1",
    "requests>=2.31.0",
    "loguru>=0.7.3",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "langchain-ollama",
    "langgraph",
    "arxiv",
    "hello-agents>=0.2.8",
    "tavily-python>=0.5.0",
    "openai>=1.12.0",
    "ddgs>=9.6.1",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = ["E", "F", "I", "D", "D401", "T201", "UP"]
lint.ignore = ["UP006", "UP007", "UP035", "D417", "E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = ["ruff>=0.12.7"]